🔖 chore: Alipay adds WAP payment (#259)
增加wap网页支付方式；优化代码结构，把支付方式处理单独拆分导`client.go`；
